+++ changelog.md
@@ -2,6 +2,7 @@
 
 ### Master
 
+* Do not delete comment written from user have same userid for danger - kwonoj
 * Fix link to `jest` in getting started docs - palleas
 * Fix yarn install instruction in getting started docs - palleas